Assign global app, move job titles to share properties, space formatting
Since the bulk of the work is in main.js, let's move the titles to
somewhere more global and spacey. In addition, let's put each one on its
own new line since we won't access this area often. New developers can
easily change/remove the items in it with the formatting change. Other
global properties which might be shared between controllers can be
placed in the sharedProperties service as well.